refactor: TestDataFactory 중복 초기화 제거 및 네이밍 변경
- 결제저장소커서페이징Test에 PersistenceTestDataFactory 필드 추가
